@@ -20,8 +20,8 @@ Gem::Specification.new do |s|
   s.add_dependency 'ansi_stream', '~> 0.0.6'
   s.add_dependency 'attr_encrypted', '~> 3.1.0'
   s.add_dependency 'autoprefixer-rails', '~> 6.4.1'
-  s.add_dependency 'coffee-rails', '~> 4.2'
-  s.add_dependency 'explicit-parameters', '~> 0.3.0'
+  s.add_dependency 'coffee-rails', '~> 5.0'
+  s.add_dependency 'explicit-parameters', '~> 0.4.0'
   s.add_dependency 'faraday', '~> 0.12.2'
   s.add_dependency 'faraday-http-cache', '~> 1.2.2'
   s.add_dependency 'gemoji', '~> 2.1'
@@ -29,18 +29,18 @@ Gem::Specification.new do |s|
   s.add_dependency 'lodash-rails', '~> 4.6.1'
   s.add_dependency 'octokit', '~> 4.12.0'
   s.add_dependency 'omniauth-github', '~> 1.3.0'
-  s.add_dependency 'pubsubstub', '~> 0.1.2'
-  s.add_dependency 'rails', '~> 5.2.1'
+  s.add_dependency 'pubsubstub', '~> 0.2.0'
+  s.add_dependency 'rails', '~> 6.0.0'
   s.add_dependency 'rails-timeago', '~> 2.13.0'
   s.add_dependency 'rails_autolink', '~> 1.1.6'
   s.add_dependency 'rake', '~> 10.0'
-  s.add_dependency 'redis-namespace', '~> 1.5.2'
-  s.add_dependency 'redis-objects', '~> 1.2.1'
-  s.add_dependency 'responders', '~> 2.4.0'
+  s.add_dependency 'redis-namespace', '~> 1.6.0'
+  s.add_dependency 'redis-objects', '~> 1.4.3'
+  s.add_dependency 'responders', '~> 3.0'
   s.add_dependency 'safe_yaml', '~> 1.0.4'
   s.add_dependency 'sass-rails', '~> 5.0'
   s.add_dependency 'securecompare', '~> 1.0.0'
   s.add_dependency 'sprockets-rails', '>= 2.3.2'
-  s.add_dependency 'state_machines-activerecord', '~> 0.5.0'
+  s.add_dependency 'state_machines-activerecord', '~> 0.6.0'
   s.add_dependency 'validate_url', '~> 1.0.0'
 end
